[Hotfix-14-DTO-계층-변경-Little-SNS-MS] 프로덕션코드
