+++ little-SNS-MS/src/main/java/com/littleSNSMS/domain/service/PostService.java
@@ -2,7 +2,8 @@ package com.littleSNSMS.domain.service;
 
 import com.littleSNSMS.domain.Post;
 import com.littleSNSMS.domain.PostRepository;
-import com.littleSNSMS.domain.dto.PostDTO;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Mono;
@@ -13,7 +14,20 @@ public class PostService {
 
     private final PostRepository postRepository;
 
-    public Mono<Long> post(PostDTO.Create dto) {
+    public Mono<Long> post(Create dto) {
         return postRepository.save(Post.post(dto));
+    }
+
+    @Getter
+    @NoArgsConstructor(force = true)
+    @RequiredArgsConstructor
+    public static final class Create {
+        private final String content;
+        private final String ownerId;
+        private final String ownerEmail;
+
+        public static Create of(String content, String ownerId, String ownerEmail) {
+            return new Create(content, ownerId, ownerEmail);
+        }
     }
 }
